backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for fetching PubMed papers"
authors = ["Your Name <you@example.com>"]  # Changed to a list of strings
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"  # Specifies Python version constraint
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
pandas = "^2.2.3"
biopython = "^1.84"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"